+++ __mocks__/utils/discordProfile.js
@@ -1,0 +1,3 @@
+module.exports = {
+  fetchDiscordProfileInfo: jest.fn(),
+};